# Improve perf of SanitizeKey